backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=2.0.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
